requirements.txt
  django                                                                  
  djangorestframework
  djangorestframework-simplejwt
  psycopg2-binary
  python-dotenv
  djangorestframework-gis

  ruff

  geopandas
  requests